requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
